Fixes warnings, removes trailing whitespace, adds @Override annotations.

+++ src/org/jitsi/impl/neomedia/codec/audio/g729/Ld8k.java
@@ -49,26 +49,26 @@ class Ld8k
 
 static final float PI =              3.14159265358979323846f;
 /**
- * Largest floating point number 
+ * Largest floating point number
  */
-static final float FLT_MAX_G729 =         1.e38f;   
+static final float FLT_MAX_G729 =         1.e38f;
 /**
  * Largest floating point number
  */
-static final float FLT_MIN_G729 =         -FLT_MAX_G729;    
+static final float FLT_MIN_G729 =         -FLT_MAX_G729;
 
 /**
- * Total size of speech buffer 
+ * Total size of speech buffer
  */
-static final int L_TOTAL =         240;   
+static final int L_TOTAL =         240;
 /**
  * LPC update frame size
  */
-static final int L_FRAME =         80;     
+static final int L_FRAME =         80;
 /**
- * Sub-frame size 
+ * Sub-frame size
  */
-static final int L_SUBFR =         40;   
+static final int L_SUBFR =         40;
 
 /*---------------------------------------------------------------------------*
  * Constants for bitstream packing                                           *
@@ -80,7 +80,7 @@ static final short BIT_1 =     0x0081;
 /**
  * Definition of zero-bit in bit-stream.
  */
-static final short BIT_0 =     0x007f; 
+static final short BIT_0 =     0x007f;
 /**
  * Definition of frame erasure flag.
  */
@@ -88,53 +88,53 @@ static final short SYNC_WORD = 0x6b21;
 /**
  * Size of bitstream frame.
  */
-static final short SIZE_WORD =       80; 
+static final short SIZE_WORD =       80;
 /**
- * Number of parameters per 10 ms frame. 
+ * Number of parameters per 10 ms frame.
  */
-static final int PRM_SIZE =        11;     
+static final int PRM_SIZE =        11;
 /**
  * Bits per frame.
  */
-static final int SERIAL_SIZE =     82;     
+static final int SERIAL_SIZE =     82;
 
 /*---------------------------------------------------------------------------*
  * Constants for lpc analysis and lsp quantizer.                             *
  *---------------------------------------------------------------------------*/
 /**
- * LPC analysis window size.   
+ * LPC analysis window size.
  */
-static final int L_WINDOW =        240;    
+static final int L_WINDOW =        240;
 /**
  * Samples of next frame needed for LPC ana.
  */
-static final int L_NEXT =          40;     
+static final int L_NEXT =          40;
 
 /**
  * LPC order.
  */
-static final int M =               10;     
+static final int M =               10;
 /**
- * LPC order+1. 
+ * LPC order+1.
  */
-static final int MP1 =            (M+1); 
+static final int MP1 =            (M+1);
 /**
  * Resolution of lsp search.
  */
-static final int GRID_POINTS =     60;     
+static final int GRID_POINTS =     60;
 
 /**
  * MA prediction order for LSP.
  */
-static final int MA_NP =           4;     
+static final int MA_NP =           4;
 /**
  * Number of modes for MA prediction.
  */
-static final int MODE =            2;      
+static final int MODE =            2;
 /**
  * Number of bits in first stage.
  */
-static final int NC0_B =           7;    
+static final int NC0_B =           7;
 /**
  * Number of entries in first stage.
  */
@@ -142,29 +142,29 @@ static final int NC0 =          (1<<NC0_B);
 /**
  * Number of bits in second stage.
  */
-static final int NC1_B =           5;       
+static final int NC1_B =           5;
 /**
  * Number of entries in second stage.
  */
-static final int NC1 =          (1<<NC1_B); 
+static final int NC1 =          (1<<NC1_B);
 /**
  * LPC order / 2.
  */
-static final int NC =              (M/2);   
+static final int NC =              (M/2);
 
-static final float L_LIMIT =         0.005f;   
-static final float M_LIMIT =         3.135f;   
-static final float GAP1 =            0.0012f; 
-static final float GAP2 =            0.0006f;  
-static final float GAP3 =            0.0392f;  
+static final float L_LIMIT =         0.005f;
+static final float M_LIMIT =         3.135f;
+static final float GAP1 =            0.0012f;
+static final float GAP2 =            0.0006f;
+static final float GAP3 =            0.0392f;
 /**
- * pi*0.04 
+ * pi*0.04
  */
-static final float PI04 =            PI*0.04f;   
+static final float PI04 =            PI*0.04f;
 /**
- * pi*0.92 
+ * pi*0.92
  */
-static final float PI92 =            PI*0.92f;   
+static final float PI92 =            PI*0.92f;
 static final float CONST12 =         1.2f;
 
 /*-------------------------------------------------------------------------
@@ -191,11 +191,11 @@ static final float BETA =        1.0f;
 /**
  * Minimum pitch lag in samples
  */
-static final int PIT_MIN =         20;     
+static final int PIT_MIN =         20;
 /**
  * Maximum pitch lag in samples
  */
-static final int PIT_MAX =         143;    
+static final int PIT_MAX =         143;
 /**
  * Length of filter for interpolation.
  */
@@ -203,23 +203,23 @@ static final int L_INTERPOL =      (10+1);
 /**
  * Length for pitch interpolation
  */
-static final int L_INTER10 =       10;   
+static final int L_INTER10 =       10;
 /**
- * upsampling ration for pitch search  
+ * upsampling ration for pitch search
  */
-static final int L_INTER4 =        4;   
+static final int L_INTER4 =        4;
 /**
  * resolution of fractional delays
  */
-static final int UP_SAMP =         3;      
+static final int UP_SAMP =         3;
 /**
- * Threshold to favor smaller pitch lags 
+ * Threshold to favor smaller pitch lags
  */
-static final float THRESHPIT =    0.85f;  
+static final float THRESHPIT =    0.85f;
 /**
- * maximum adaptive codebook gain 
+ * maximum adaptive codebook gain
  */
-static final float GAIN_PIT_MAX = 1.2f;     
+static final float GAIN_PIT_MAX = 1.2f;
 static final int FIR_SIZE_ANA = (UP_SAMP*L_INTER4+1);
 static final int FIR_SIZE_SYN = (UP_SAMP*L_INTER10+1);
 
@@ -227,29 +227,29 @@ static final int FIR_SIZE_SYN = (UP_SAMP*L_INTER10+1);
  * Constants for fixed codebook.                                            *
  *---------------------------------------------------------------------------*/
 /**
- * Size of correlation matrix  
+ * Size of correlation matrix
  */
-static final int DIM_RR =  616; 
+static final int DIM_RR =  616;
 /**
- * Number of positions for each pulse 
+ * Number of positions for each pulse
  */
-static final int NB_POS =  8;   
+static final int NB_POS =  8;
 /**
  * Step betweem position of the same pulse.
  */
-static final int STEP =    5;  
+static final int STEP =    5;
 /**
  * Size of vectors for cross-correlation between 2 pulses
  */
-static final int MSIZE =   64;  
+static final int MSIZE =   64;
 /**
  * Maximum value of pitch sharpening
  */
-static final float SHARPMAX =        0.7945f;  
+static final float SHARPMAX =        0.7945f;
 /**
  * minimum value of pitch sharpening
  */
-static final float SHARPMIN =        0.2f;    
+static final float SHARPMIN =        0.2f;
 
 /*--------------------------------------------------------------------------*
  * Example values for threshold and approximated worst case complexity:     *
@@ -257,7 +257,7 @@ static final float SHARPMIN =        0.2f;
  *     threshold=0.40   maxtime= 75   extra=30   Mips =  6.0                *
  *--------------------------------------------------------------------------*/
 static final float THRESHFCB =       0.40f;
-static final int MAX_TIME =        75;      
+static final int MAX_TIME =        75;
 
 /*--------------------------------------------------------------------------*
  * Constants for taming procedure.                           *
@@ -265,20 +265,20 @@ static final int MAX_TIME =        75;
 /**
  * Maximum pitch gain if taming is needed
  */
-static final float GPCLIP =      0.95f;   
+static final float GPCLIP =      0.95f;
 /**
  * Maximum pitch gain if taming is needed
  */
-static final float GPCLIP2 =     0.94f;   
+static final float GPCLIP2 =     0.94f;
 /**
- * Maximum pitch gain if taming is needed 
+ * Maximum pitch gain if taming is needed
  */
 static final float GP0999 =      0.9999f;
 /**
  * Error threshold taming
  */
-static final float THRESH_ERR =  60000.0f;  
-static final float INV_L_SUBFR = (1.0f/(float)L_SUBFR); /* =0.025 */
+static final float THRESH_ERR =  60000.0f;
+static final float INV_L_SUBFR = (1.0f/L_SUBFR); /* =0.025 */
 /*-------------------------------------------------------------------------
  * gain quantizer  constants
  *-------------------------------------------------------------------------
@@ -286,31 +286,31 @@ static final float INV_L_SUBFR = (1.0f/(float)L_SUBFR); /* =0.025 */
 /**
  * Average innovation energy
  */
-static final float MEAN_ENER =        36.0f;   
+static final float MEAN_ENER =        36.0f;
 /**
- * Number of Codebook-bit  
+ * Number of Codebook-bit
  */
-static final int NCODE1_B =  3;               
+static final int NCODE1_B =  3;
 /**
- * Number of Codebook-bit      
+ * Number of Codebook-bit
  */
-static final int NCODE2_B =  4;               
+static final int NCODE2_B =  4;
 /**
- * Codebook 1 size     
+ * Codebook 1 size
  */
-static final int NCODE1 =    (1<<NCODE1_B);   
+static final int NCODE1 =    (1<<NCODE1_B);
 /**
  * Codebook 2 size
  */
-static final int NCODE2 =    (1<<NCODE2_B);   
+static final int NCODE2 =    (1<<NCODE2_B);
 /**
  * Pre-selecting order for #1
  */
-static final int NCAN1 =            4;      
+static final int NCAN1 =            4;
 /**
  * Pre-selecting order for #2
  */
-static final int NCAN2 =            8;     
+static final int NCAN2 =            8;
 static final float INV_COEF =   -0.032623f;
 
 /*---------------------------------------------------------------------------
@@ -318,26 +318,26 @@ static final float INV_COEF =   -0.032623f;
  *---------------------------------------------------------------------------
  */
 /* short term pst parameters :  */
-/** 
+/**
  * denominator weighting factor
  */
-static final float GAMMA1_PST =      0.7f;  
+static final float GAMMA1_PST =      0.7f;
 /**
  * numerator  weighting factor
  */
-static final float GAMMA2_PST =      0.55f;    
-/** 
+static final float GAMMA2_PST =      0.55f;
+/**
  * impulse response length
  */
-static final int LONG_H_ST =       20;     
-/** 
+static final int LONG_H_ST =       20;
+/**
  * tilt weighting factor when k1>0
  */
-static final float GAMMA3_PLUS =     0.2f;     
-/** 
+static final float GAMMA3_PLUS =     0.2f;
+/**
  * tilt weighting factor when k1<0
  */
-static final float GAMMA3_MINUS =    0.9f;    
+static final float GAMMA3_MINUS =    0.9f;
 
 /* long term pst parameters :   */
 /**
@@ -345,40 +345,40 @@ static final float GAMMA3_MINUS =    0.9f;
  */
 static final int L_SUBFRP1 = (L_SUBFR + 1);
 /**
- * resolution for fractionnal delay   
+ * resolution for fractionnal delay
  */
-static final int F_UP_PST =        8;       
+static final int F_UP_PST =        8;
 /**
- * length of short interp. subfilters  
+ * length of short interp. subfilters
  */
-static final int LH2_S =           4;       
+static final int LH2_S =           4;
 /**
  * length of long interp. subfilters
  */
-static final int LH2_L =           16;      
+static final int LH2_L =           16;
 /**
- * threshold LT pst switch off       
+ * threshold LT pst switch off
  */
-static final float THRESCRIT =       0.5f;     
+static final float THRESCRIT =       0.5f;
 /**
- * LT weighting factor          
+ * LT weighting factor
  */
-static final float GAMMA_G =         0.5f;     
+static final float GAMMA_G =         0.5f;
 /**
- * gain adjustment factor             
+ * gain adjustment factor
  */
-static final float AGC_FAC =         0.9875f; 
+static final float AGC_FAC =         0.9875f;
 /**
- * gain adjustment factor                 
+ * gain adjustment factor
  */
-static final float AGC_FAC1 =         (1.f - AGC_FAC);    
+static final float AGC_FAC1 =         (1.f - AGC_FAC);
 static final int LH_UP_S =         (LH2_S/2);
 static final int LH_UP_L =         (LH2_L/2);
 static final int LH2_L_P1 =    (LH2_L + 1);
 /**
- * LT gain minimum         
+ * LT gain minimum
  */
-static final float MIN_GPLT =     (1.f / (1.f + GAMMA_G));  
+static final float MIN_GPLT =     (1.f / (1.f + GAMMA_G));
 
 /* Array sizes */
 static final int MEM_RES2 = (PIT_MAX + 1 + LH_UP_L);
